python/requirements.txt
pyserial
protobuf
nanopb
cobs